Create indexes for id values bag tables.
The bag tables are  loaded without primary key.
Adding the primary key wit ALTER statements in with
docker-compose fails becauise the statement cannot be passed.

There is no eays way to add primary keys if they do not exist
Therefore we add a index instead of the primary key to get the same
performance benefits.